Begin switching to vue files. Implement App.vue
