+++ src/client/DroppedPage.vue
