+++ src/client/index.html
@@ -2,6 +2,7 @@
 <html>
 	<head>
 		<meta charset="UTF-8" />
+		<meta http-equiv="Content-Security-Policy" content="default-src 'none'; script-src 'unsafe-inline' 'self'; style-src 'self' https://maxcdn.bootstrapcdn.com https://use.fontawesome.com; font-src https://use.fontawesome.com;">
 		<title>CheatSheet</title>
 
 		<!-- Bootstrap CSS -->
@@ -27,7 +28,7 @@
 
 		<!-- Custom files -->
 		<link rel="stylesheet" type="text/css" href="style.css" />
-		<script src="App.js"></script>
+		<script src="dist/main.js" type="text/javascript" defer></script>
 
 		<!-- Open all links in the default browser -->
 		<script>
